% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{calc_save_climatology}
\alias{calc_save_climatology}
\title{Calculate climatology}
\usage{
calc_save_climatology(path.to.file, file.to.write)
}
\arguments{
\item{path.to.file}{all the full names of the files to stack}

\item{file.to.write}{name of the climatology file to save}
}
\description{
calculate the climatology for the entire run period, uses \link[GeoInterpolation]{load.data.paths}. 
See also \link[GeoInterpolation]{plot_climatology}
}
\details{
Calculate climatology
}
\author{
Marieke Dirksen
}
