% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coords_aws}
\alias{coords_aws}
\title{Coordinates of the AWS station in RD}
\format{A data frame with 50 rows and 8 variables:
\describe{
 \item{STN}{station number}
 \item{DS_ALT}{altitude [m]}
 \item{DS_NAME}{name site}
 \item{DS_LAT}{latitude WGS84}
 \item{DS_LON}{longitude WGS84}
 \item{optional}{TRUE or FALSE}
 \item{RDN_X}{rijksdriehoeks x-coordinates}
 \item{RDN_Y}{rijksdriehoeks y-coordinates}
}}
\usage{
coords_aws
}
\description{
Description of the sites, inlcuding coordinates in lat lon and Rijksdriehoekstelsel for all AWS stations
}
\keyword{datasets}
