% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossvalidate.r
\name{createCvObjFromSpatialObj}
\alias{createCvObjFromSpatialObj}
\title{create a object for cross validation}
\usage{
createCvObjFromSpatialObj(obj)
}
\arguments{
\item{obj}{a object}
}
\description{
creating a cross validation object from spatial object
}
