% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distsea.grid}
\alias{distsea.grid}
\title{Distance to the sea grid}
\format{An object of class \code{SpatialGridDataFrame} with 84372 rows and 1 columns.}
\usage{
distsea.grid
}
\description{
In order to estimate the distance to the sea within the Netherlands a virtual shoreline is drawn along the Dutch coast.
The "shoreline" is North of the Waddenzee.
}
\section{Slots}{

\describe{
\item{\code{distsea}}{distance to the virtual shoreline in meters}
}}

\keyword{datasets}
