% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doNearestNeighbor.r
\name{doNearestNeighbor}
\alias{doNearestNeighbor}
\title{Performs nearest neighbor interpolation}
\usage{
doNearestNeighbor(formula, data, newdata, debug.level = 1)
}
\arguments{
\item{formula}{A formula to define the dependent and the independent variables, see the documentation of krige. 
Note that nearest neighbor does not support independent variables. Therefor, the formula should always 
have the form dependent~1.}

\item{data}{The input data, should be a spatial object which supports coordinates extracting through coordinates()}

\item{newdata}{A spatial object with the prediction locations.}

\item{debug.level}{standard value set to 1}

\item{...}{parameters that are passed on to Tps}
}
\value{
The function returns a spatial object with the same class as \code{newdata} with the prediction and the 
         variance (NA in this case). The names of the columns match the outcome of the krige function.
}
\description{
Performs nearest neighbor interpolation
}
\note{
This functions uses idw with 'nmax' set to 1 to perform nearest neighbor interpolation.
}
\examples{
library(sp)
data(meuse)
coordinates(meuse) = ~x+y
data(meuse.grid)
gridded(meuse.grid) = ~x+y

meuse_nn = doNearestNeighbor(zinc~1, meuse, meuse.grid)

}
\author{
Paul Hiemstra, \email{p.h.hiemstra@gmail.com}
}
