% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doTps.r
\name{doTps}
\alias{doTps}
\title{Provides a krige (automap, gstat) like interface to Tps (fields)}
\usage{
doTps(formula, data, newdata, ..., debug.level = 1, addFit = FALSE)
}
\arguments{
\item{formula}{A formula to define the dependent and the independent variables, see the documentation of krige.}

\item{data}{The input data, should be a spatial object which supports coordinates extracting through coordinates()}

\item{newdata}{A spatial object with the prediction locations.}

\item{...}{parameters that are passed on to Tps}

\item{debug.level}{standard value set to 1}

\item{addFit}{logical, whether or not the output contains the fitted Tps model. See the value section below for more details.}
}
\value{
a spatial object with the same class as \code{newdata} with the prediction and the variance (NA in this case).
        The names of the columns match the outcome of the krige function. When \code{addFit} equals \code{TRUE}
        the function returns an autoKrige object with two elements: \code{krige_output} with the Tps prediction and variance
        and \code{fit} with the fitted Tps model.
}
\description{
Provides a krige (automap, gstat) like interface to Tps (fields)
}
\examples{
library(sp)
data(meuse)
coordinates(meuse) = ~x+y
data(meuse.grid)
gridded(meuse.grid) = ~x+y

meuse_tps = doTps(zinc~dist, meuse, meuse.grid)
meuse_tps = doTps(zinc~1, meuse, meuse.grid)

summary(meuse_tps$krige_output)
meuse_tps$fit
}
\author{
Paul Hiemstra, \email{p.h.hiemstra@gmail.com}
}
