% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadWunderground.R
\name{download_time_seq}
\alias{download_time_seq}
\title{Wunderground time series from URL}
\usage{
download_time_seq(stationname = "IZUIDHOL87",
  start = as.Date("2015/03/01"), stop = as.Date("2018/10/01"),
  var = "TemperatureC",
  folder.loc = "/nobackup/users/dirksen/data/wunderground/")
}
\arguments{
\item{stationname}{Character, name of the station}

\item{start}{Date, start of the download period}

\item{stop}{Date, stop period of the download period}

\item{var}{Character, variable(s) which you want to download}

\item{folder.loc}{Character, location were the data will be stored. For each stationname a new folder is created.}
}
\description{
Function to download 5min weather data from the \url{www.wunderground.nl} website. For a specific time period, 
station and variable(s) the data can be downloaded and saved in a directory. For each stationname a new folder is created (unless)
the folder already exists. After downloading the daily files all the .txt files are combined and the daily files are deleted.
}
\details{
Download Wunderground data
}
\examples{
#Run the example below to store data at the current location
download_time_seq(stationname = "IZUIDHOL87",
                  start = as.Date("2018/04/01"),
                  stop = as.Date("2018/05/01"),
                  var = "TemperatureC",
                  folder.loc = ".")
For multiple stations use lapply:
stations<-c('IZUIDHOL36','ILIMBURG36')
lapply(stations,download_time_seq)
}
