% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_statistical_summary.R
\name{get_statistical_summary}
\alias{get_statistical_summary}
\title{Statistics from interpolation}
\usage{
get_statistical_summary(observed, prediction)
}
\arguments{
\item{observed}{observation from your groundstations}

\item{prediction}{prediction at the location of the groundstations (either from the cv or prediction)}
}
\description{
makes a statistical summary list from the output of interpolation kriging
}
\details{
look at \code{interpolation_kriging}
}
\author{
Marieke Dirksen
}
