% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation_idw.R
\name{interpolation_idw}
\alias{interpolation_idw}
\title{IDW interpolation}
\usage{
interpolation_idw(groundstations, variable, distshore.grd,
  mxdidw = 120000, blocksize = 20000)
}
\arguments{
\item{groundstations}{Data from the ground station in spatial format without na values}

\item{variable}{the variable of the groundstations you want to interpolate}

\item{distshore.grd}{masking your output for a certain region, default is the Netherlands}

\item{mxdidw}{idw interpolation variable, with a standard value of 120000}

\item{blocksize}{idw interpolation variable, with a standard value of 20000}
}
\value{
returns a list with the kriging interpolation output and a dataframe with the difference between your observations and predicted grid
}
\description{
Function overlays the ground stations and the grid and interpolates using idw.
}
\details{
after running this function take a look
}
\author{
Marieke Dirksen
}
