% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation_krigingdrift.R
\name{interpolation_krigingdrift}
\alias{interpolation_krigingdrift}
\title{Kriging with External Drift interpolation}
\usage{
interpolation_krigingdrift(groundstations, method.overlay = "extract",
  formula, variable, grid_drift, blocksize = 20000, mxdkrige = Inf,
  mod = "Exp", conditional.sim = FALSE)
}
\arguments{
\item{groundstations}{Data from the ground station in spatial format without na values}

\item{method.overlay}{Method to overlay the points and the grid, choose between over (1 cell) and extract function (which averages over 4 nearest cells)}

\item{formula}{formula used for the kriging interpolation}

\item{variable}{the variable of the groundstations you want to interpolate}

\item{grid_drift}{trend grid for the interpolation with same spatial extent as groundstations}

\item{blocksize}{kriging interpolation variable, with a standard value of 20000}

\item{mxdkrige}{kriging interpolation variable, with a standard value of Inf}

\item{mod}{kriging interpolation variable, standard set to "Exp"}

\item{conditional.sim}{boolean, TRUE or FALSE. Standard set to FALSE as this simulation takes more time, nmax=30 and nsim=4.}
}
\value{
returns a list with the kriging interpolation output and a dataframe with the difference between your observations and predicted grid
}
\description{
Function overlays the ground stations and the grid and interpolates using autokrige from the \code{\link{autoKrige}}.
}
\details{
after running this function take a look at \code{get_statistical_summary}
}
\author{
Marieke Dirksen
}
