% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation_ordinarykriging.R
\name{interpolation_ordinarykriging}
\alias{interpolation_ordinarykriging}
\title{Ordinary Kriging interpolation}
\usage{
interpolation_ordinarykriging(groundstations, variable, grid_drift,
  blocksize = 20000, mxdkrige = Inf, mod = "Exp")
}
\arguments{
\item{groundstations}{Data from the ground station in spatial format without na values}

\item{variable}{the variable of the groundstations you want to interpolate}

\item{grid_drift}{trend grid for the interpolation with same spatial extent as groundstations}

\item{blocksize}{kriging interpolation variable, with a standard value of 20000}

\item{mxdkrige}{kriging interpolation variable, with a standard value of Inf}

\item{mod}{kriging interpolation variable, standard set to "Exp"}
}
\value{
returns a list with the kriging interpolation output and a dataframe with the difference between your observations and predicted grid
}
\description{
Function overlays the ground stations and the grid and interpolates using autokrige from the \code{\link{autoKrige}}.
}
\details{
after running this function take a look at \code{get_statistical_summary}
}
\author{
Marieke Dirksen
}
