% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation_tps.R
\name{interpolation_tps}
\alias{interpolation_tps}
\title{IDW interpolation}
\usage{
interpolation_tps(groundstations, variable, distshore.grd)
}
\arguments{
\item{groundstations}{Data from the ground station in spatial format without na values}

\item{variable}{the variable of the groundstations you want to interpolate}

\item{distshore.grd}{masking your output for a certain region, default is the Netherlands}
}
\value{
returns a list with the kriging interpolation output and a dataframe with the difference between your observations and predicted grid
}
\description{
Function overlays the ground stations and the grid and interpolates using tps.
}
\details{
after running this function take a look at \code{get_statistical_summary}
}
\author{
Marieke Dirksen
}
