% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{load.data.paths}
\alias{load.data.paths}
\title{Load data paths}
\usage{
load.data.paths(results.aux = results.aux, results.pca = results.pca)
}
\arguments{
\item{results.aux}{results from the auxiliary data runs}

\item{results.pca}{results from the principle component runs}
}
\description{
Loading the interpolation results file names into memory. See also 
\link[GeoInterpolation]{calc_save_climatology} and \link[GeoInterpolation]{plot_climatology}
}
\details{
Load data paths
}
\author{
Marieke Dirksen
}
