% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_auxiliarydata.R
\name{load_auxiliarydata}
\alias{load_auxiliarydata}
\title{load auxiliary data}
\usage{
load_auxiliarydata(datum)
}
\arguments{
\item{datum}{datum in the format Ymd}
}
\value{
returns a raster stack with all the layers
}
\description{
find and loads all the auxiliary data used for the interpolation and returns a raster stack.
}
\details{
sarah insolation, siccs data, albedo from modis, snow coverage, snow albedo modis, distance to the sea,
ndvi, precipitation and roughness are all loaded.
}
\author{
Marieke Dirksen
}
