% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{load_statistical_summary}
\alias{load_statistical_summary}
\title{load_statistical_summary}
\usage{
load_statistical_summary(results.aux = results.aux,
  results.pca = results.pca)
}
\arguments{
\item{results.aux}{location were the text files are}

\item{results.pca}{location were the text files from the PCA analysis are}
}
\description{
Loads all de text files with RMSE and R2 values into memory.
}
\details{
Loading the statistical summary files
}
\author{
Marieke Dirksen
}
