% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_variance.R
\name{pca_variance}
\alias{pca_variance}
\title{pca variance}
\usage{
pca_variance(sdev)
}
\arguments{
\item{sdev}{standard deviation of the PCA}
}
\description{
\code{calculates variance from the sandard deviation of a principle component analysis}
}
\author{
Marieke Dirksen
}
