% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{plot_climatology}
\alias{plot_climatology}
\title{plot climatology}
\usage{
plot_climatology(list.of.files, names.attr, layout, at = seq(9.5, 12.3,
  length = 30))
}
\arguments{
\item{list.of.files}{climatologies (grd)}

\item{names.attr}{names of the different subplots}

\item{layout}{how should the plots be ordered?}
}
\description{
plots the climatology calculated from \link[GeoInterpolation]{calc_save_climatology}. 
See also \link[GeoInterpolation]{load.data.paths}
}
\details{
Plotting climatology
}
\author{
Marieke Dirksen
}
