% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_over_raster.R
\name{points_over_raster}
\alias{points_over_raster}
\title{Points over raster}
\usage{
points_over_raster(datum, spdf)
}
\arguments{
\item{datum}{date in the format YYYY-MM-DD}

\item{spdf}{spatial pixel dataframe}
}
\value{
returns a dataframe with all auxiliary data and the datum
}
\description{
Overlay spatial points over the auxiliary datasets for a specific day. Uses \code{\link{load_auxiliarydata}}
}
\examples{
datum<-"1991-05-14"
data("coords_aws")

coordinates(coords_aws)<-~RDN_X+RDN_Y
plot(coords_aws)

aux.data<-points_over_raster(datum = datum, spdf = coords_aws)
}
