% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOM.R
\name{rasterSOM}
\alias{rasterSOM}
\title{Self Organizing Maps for rasters}
\usage{
rasterSOM(x, xdim = 3, ydim = 5, rlen = 1000, alpha = 0.2,
  datum = NULL)
}
\arguments{
\item{x}{Raster* format}

\item{xdim}{x-dimension of the map nodes}

\item{ydim}{y-dimension of the map nodes.}

\item{rlen}{number of iterations to fit the SOMs}

\item{alpha}{learning rate}

\item{datum}{optional. Used to calculate the relative frequency of each SOM for each month.}
}
\description{
Self organizing maps for raster datasets. The function uses the som function from the kohonen package to calculate the 
SOMs. The map nodes are typically 3x4 or 6x4. The update radius is based on the input data. The learning rate should decrease with the number of iterations.
Richardson(2003) used the following settings: 

\itemize{
\item 1000 iterations with learning rate of 0.2
\item 10.000 iterations with a learning rate of 0.1
\item 50.000 iterations with a learning rate of 0.002
}
}
\examples{
#not run
#plot(som_model,type="changes")
}
\author{
Marieke Dirksen
}
