% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SICCS_preprocess.R
\name{SICCS_preprocess}
\alias{SICCS_preprocess}
\title{SICCS preprocess}
\usage{
SICCS_preprocess(file_loc, crs1 = NA, crs2 = NA, new_grid,
  write_raster = FALSE, save_dir = NULL, file.format = "h5")
}
\arguments{
\item{file_loc}{location of the h5 files}

\item{crs1}{original crs}

\item{crs2}{crs to which the file needs to be converted}

\item{new_grid}{grid to which the file will be cropped, resampled and masked}

\item{write_raster}{TRUE or FALSE, standard set to FALSE}

\item{save_dir}{only required if write_raster=TRUE, file name should end with ".grd"}

\item{file.format}{"h5" and "hdf5" both accepted}
}
\description{
Converts the h5 files from SICCS to grid files
}
\details{
Preprocess SICCS files
}
\author{
Marieke Dirksen
}
