% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{solar_radiation}
\alias{solar_radiation}
\title{Radiation data from stations}
\format{A data frame with 145063 rows and 7 variables:
\describe{
 \item{Datum}{Day of the measurement in YYYY-MM-DD format}
 \item{DS_CODE}{unique code name of the observation station 3 numbers ending with _H}
 \item{RDN_X}{x-coordinates of the station in the Dutch RD-coordinate system}
 \item{RDN_Y}{y-coordinates of the station in the Dutch RD-coordinate system}
 \item{Q}{Incomming Radiation from the ground-based observations}
 \item{SAT}{Incomming Radiation from the satellite observations}
 \item{DIFF}{Difference between the ground-based observations and satellite observations}
}}
\usage{
solar_radiation
}
\description{
This data-set contains KNMI incomming radiation data from 32 stations and satellite observations from MSG at these locations, from 2004 until 2016.
The satellite product SICCS was used and compared using the 'over' function.
}
\keyword{datasets}
