% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temperature_climate}
\alias{temperature_climate}
\title{Climatological record mean temperature data from AWS stations}
\format{A data frame with 635814 rows and 3 variables:
\describe{
 \item{STN}{station number}
 \item{Datum}{Day of the measurement in YYYY-MM-DD format}
 \item{Tg}{Mean temperature in degrees celcius}
}}
\usage{
temperature_climate
}
\description{
This data-set contains KNMI temperature data from AWS stations, from 1901 until 2017. The time series is homogenious.
}
\keyword{datasets}
