% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temperature_min_max}
\alias{temperature_min_max}
\title{Daily Minimum and Maximum temperatures dataset}
\format{A data frame with 325411 rows and 6 variables:
\describe{
 \item{Datum}{Datum in the format YYYY-MM-DD}
 \item{DS_CODE}{unique code name of the observation station 3 numbers ending with _H}
 \item{STN}{code name of the observation station 3 numbers}
 \item{TN}{Minimum temperature of the day out of the 6 hour minimum temperatures}
 \item{TX}{Maximum temperature of the day out of the 6 hour maximum temperatures}
 \item{DTR}{Diurnal Temperature Range (DTR) calculated as TX-TN}
}}
\usage{
temperature_min_max
}
\description{
This data-set contains KNMI temperature data from 34 stations, from 1990 until 2017.
}
\keyword{datasets}
