% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temperature_stationdata}
\alias{temperature_stationdata}
\title{Temperature data from stations}
\format{A data frame with 229183 rows and 8 variables:
\describe{
 \item{Stn}{station number}
 \item{Locatie}{Location name of the station}
 \item{RDN_X}{x-coordinates of the station in the Dutch RD-coordinate system}
 \item{RDN_Y}{y-coordinates of the station in the Dutch RD-coordinate system}
 \item{Datum}{Day of the measurement in YYYYMMDD format}
 \item{Tg}{Mean temperature in degrees celcius}
 \item{Tn}{Minimum temperature in degrees celcius}
 \item{Tx}{Maximum temperature in degrees celcius}
}}
\usage{
temperature_stationdata
}
\description{
This data-set contains KNMI temperature data from 34 stations, from 1995 until 2014.
}
\keyword{datasets}
