% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wunderground_meta}
\alias{wunderground_meta}
\title{Metadata from Wunderground stations}
\format{A data frame with 10 rows and 8 variables:
\describe{
 \item{Station.ID}{station name}
 \item{Lat}{Latitude of the station in WGS84}
 \item{Lon}{Longitude of the station in WGS84}
 \item{System}{Hardware used for the meteorological measurements}
 \item{Elevation}{Elevation of the station}
 \item{start}{Date of the first measurement of the station}
 \item{stop}{Date of the last measurement}
 \item{number_obs}{Number of observations between the start and stop dates}
}}
\usage{
wunderground_meta
}
\description{
Metadata from Wunderground stations
}
\keyword{datasets}
